v6.12.0. Improvement. Bugfix.
- v6.12.0 November 23, 2012
	- When creating new documents or files, if it is inside an unknown
path we will now default to creating a document intead of a file
	- We now send growl notifications when errors occur
		- Thanks to [Luke Hagan](https://github.com/lhagan) for [pull request
#346](https://github.com/bevry/docpad/pull/346) and [issue
#343](https://github.com/bevry/docpad/issues/343)
	- We now error and provide suggestions when an extension transform
doesn't do anything
		- Thanks to [Farid Neshat](https://github.com/alFReD-NSH), [Elias
Dawson](https://github.com/eliasdawson) and [Steve
Trevathan](https://github.com/kidfribble) for [issue
#192](https://github.com/bevry/docpad/issues/192)
	- Watching stability has been improved signifcantly
		- Thanks to [ashnur](https://github.com/ashnur) for [issue
#283](https://github.com/bevry/docpad/issues/283)
	- Parser headers that don't include spacing now work again (e.g.
`---cson` instead of `--- cson`)
		- Thanks to [bobobo1618](https://github.com/bobobo1618) for [issue
#341](https://github.com/bevry/docpad/issues/341)
	- Removed default comparator on `FilesCollection` due to performance
improvement it provides
		- Thanks to [Bruno Héridet](https://github.com/Delapouite) for [issue
#330](https://github.com/bevry/docpad/issues/330)
	- Added
		- `Document::parseFileDirectory(opts,next)`
		- `Document::parseDocumentDirectory(opts,next)`
		- `FilesCollection::fuzzyFindOne(data)`

+++ src/lib/models/file.coffee
@@ -160,16 +160,20 @@ class FileModel extends Model
 			data = attrs.data
 			delete attrs.data
 			delete @attributes.data
-
-		# Apply
-		@outDirPath = outDirPath  if outDirPath
-		@setStat(stat)  if stat
-		@setData(data)  if data
-		@set({
+		defaults =
 			extensions: []
 			urls: []
 			id: @cid
-		})
+
+		# Apply
+		@outDirPath = outDirPath  if outDirPath
+		@setData(data)  if data
+		if stat
+			@setStat(stat)
+		else
+			defaults.ctime = new Date()
+			defaults.mtime = new Date()
+		@set(defaults)
 
 		# Meta
 		@meta = new Model()
@@ -534,6 +538,12 @@ class FileModel extends Model
 		content = @get('content') or @getData()
 		encoding = @get('encoding')
 
+		# Check
+		# Sometimes the out path could not be set if we are early on in the process
+		unless fileOutPath
+			next()
+			return @
+
 		# Log
 		file.log 'debug', "Writing the file: #{fileOutPath} #{encoding}"
 
@@ -558,19 +568,29 @@ class FileModel extends Model
 		file = @
 		fileOutPath = @get('outPath')
 
+		# Check
+		# Sometimes the out path could not be set if we are early on in the process
+		unless fileOutPath
+			next()
+			return @
+
 		# Log
 		file.log 'debug', "Delete the file: #{fileOutPath}"
 
-		# Write data
-		balUtil.unlink fileOutPath, (err) ->
-			# Check
-			return next(err)  if err
+		# Check existance
+		balUtil.exists fileOutPath, (exists) ->
+			# Exit if it doesn't exist
+			return next()  unless exists
+			# If it does exist delete it
+			balUtil.unlink fileOutPath, (err) ->
+				# Check
+				return next(err)  if err
 
-			# Log
-			file.log 'debug', "Deleted the file: #{fileOutPath}"
+				# Log
+				file.log 'debug', "Deleted the file: #{fileOutPath}"
 
-			# Next
-			next()
+				# Next
+				next()
 
 		# Chain
 		@
